Use latest date with data hook
